+++ src/ska_tmc_dishleafnode/commands/configure_command.py
@@ -54,8 +54,8 @@ class Configure(DishLNCommand):
         )
 
         self.track_table_process = None
-        self.timeout_id = f"{time.time()}_{__class__.__name__}"
-        self.timeout_callback = TimeoutCallback(self.timeout_id, self.logger)
+        self.timeout_id = None
+        self.timeout_callback = None
         self.task_callback: Callable
         self.partial_configure = False
         self.receiver_band: str = ""
@@ -82,6 +82,9 @@ class Configure(DishLNCommand):
         :return: : None
         :rtype: None
         """
+        self.timeout_id = f"{time.time()}_{__class__.__name__}"
+        self.timeout_callback = TimeoutCallback(self.timeout_id, self.logger)
+        self.component_manager.abort_event = task_abort_event
         # Indicate that the task has started
         self.task_callback = task_callback
         self.task_callback(status=TaskStatus.IN_PROGRESS)
@@ -133,7 +136,7 @@ class Configure(DishLNCommand):
         if return_code in [ResultCode.FAILED, ResultCode.REJECTED]:
             self.logger.info(
                 "Setting taskcallback to FAILED with message: %s",
-                lrcr_callback,
+                message,
             )
             lrcr_callback(
                 self.component_manager.command_id,
@@ -162,10 +165,15 @@ class Configure(DishLNCommand):
 
             if status == TaskStatus.ABORTED:
                 self.task_callback(status=status)
-                return
-            if result[0] == ResultCode.OK:
+                self.component_manager.command_in_progress = None
+                self.logger.info("Configure command execution is aborted.")
+            elif result[0] == ResultCode.OK:
                 self.component_manager.command_in_progress = None
                 self.task_callback(result=result, status=status)
+                self.logger.info(
+                    "Executed the Configure command successfully on %s",
+                    self.dish_master_adapter.dev_name,
+                )
             else:
                 self.task_callback(
                     status=status,
@@ -192,7 +200,7 @@ class Configure(DishLNCommand):
 
         except Exception as e:
             self.logger.exception(
-                "Exception occured while updating task status %s", e
+                "Exception occurred while updating task status %s", e
             )
 
     def update_task_callback(
@@ -347,7 +355,7 @@ class Configure(DishLNCommand):
             self.component_manager.create_process_and_start_track_table_calculation()  # noqa: E501
             # pylint: enable=line-too-long
 
-            self.receiver_band = json_argument["dish"]["receiver_band"]
+            # self.receiver_band = json_argument["dish"]["receiver_band"]
 
             self.logger.info("Invoking ConfigureBand command")
 
@@ -398,12 +406,20 @@ class Configure(DishLNCommand):
 
             if self.component_manager.dishMode != DishMode.STOW:
                 result_code, message = self.ensure_dish_is_configured()
-                if result_code[0] in [ResultCode.FAILED, ResultCode.REJECTED]:
+                if result_code[0] in [
+                    ResultCode.FAILED,
+                    ResultCode.REJECTED,
+                    ResultCode.ABORTED,
+                ]:
                     return result_code[0], message[0]
 
                 result_code, message = self.start_dish_tracking(json_argument)
 
-                if result_code[0] in [ResultCode.FAILED, ResultCode.REJECTED]:
+                if result_code[0] in [
+                    ResultCode.FAILED,
+                    ResultCode.REJECTED,
+                    ResultCode.ABORTED,
+                ]:
                     return result_code[0], message[0]
 
         except Exception as exception:
@@ -513,8 +529,17 @@ class Configure(DishLNCommand):
             )
         self.component_manager.update_source_offset_callback(offsets_argin)
 
-        result: bool = self.set_wait_for_trackloadstaticoff_completed()
-        if result:
+        result: str = self.set_wait_for_trackloadstaticoff_completed()
+
+        if result == "ABORTED":
+            self.logger.info(
+                "AbortCommands() command is invoked on the DishLeafNode."
+            )
+            return (
+                ResultCode.ABORTED,
+                "AbortCommands() command is invoked on the DishLeafNode.",
+            )
+        if result == "ACHIEVED":
             if (
                 self.component_manager.track_load_static_off_result[
                     "result_code"
@@ -557,7 +582,11 @@ class Configure(DishLNCommand):
         return: Tuple[ResultCode, str]"""
         if self.component_manager.dishMode != DishMode.OPERATE:
             result_code, message = self.ensure_dish_in_right_dish_mode()
-            if result_code[0] in [ResultCode.FAILED, ResultCode.REJECTED]:
+            if result_code[0] in [
+                ResultCode.FAILED,
+                ResultCode.REJECTED,
+                ResultCode.ABORTED,
+            ]:
                 return result_code, message
 
         return self.invoke_track_command(json_argument)
@@ -572,8 +601,17 @@ class Configure(DishLNCommand):
         return: Tuple[ResultCode, str]
         """
         # Set wait for dish band to be configured
-        result: bool = self.set_wait_for_configured_band()
-        if result:
+        result: str = self.set_wait_for_configured_band()
+        if result == "ABORTED":
+            self.logger.info(
+                "AbortCommands() command is invoked on the DishLeafNode."
+            )
+            return (
+                [ResultCode.ABORTED],
+                ["AbortCommands() command is invoked on the DishLeafNode."],
+            )
+
+        if result == "ACHIEVED":
             if (
                 self.component_manager.configure_band_result["result_code"]
                 == ResultCode.FAILED
@@ -659,8 +697,16 @@ class Configure(DishLNCommand):
                 [self.component_manager.set_operate_mode_result["exception"]],
             )
 
-        result: bool = self.set_wait_for_setoperatemode_completed()
-        if result:
+        result: str = self.set_wait_for_setoperatemode_completed()
+        if result == "ABORTED":
+            self.logger.info(
+                "AbortCommands() command is invoked on the DishLeafNode."
+            )
+            return (
+                [ResultCode.ABORTED],
+                ["AbortCommands() command is invoked on the DishLeafNode."],
+            )
+        if result == "ACHIEVED":
             if (
                 self.component_manager.set_operate_mode_result["result_code"]
                 == ResultCode.FAILED
@@ -700,17 +746,23 @@ class Configure(DishLNCommand):
         :return: Resulcode and message
         :rtype: tuple
         """
-        result: bool = self.is_tracktable_provided()
-        if not result:
+        result: str = self.is_tracktable_provided()
+        if result == "ABORTED":
+            self.logger.info(
+                "Configure command has been aborted."
+                + " Track command is not invoked."
+            )
+            return ([ResultCode.ABORTED], ["Command has been aborted"])
+        if result == "FALSE":
             self.logger.error(
-                "Timed out while waiting to generate TrackTable."
-                + "Track command will not be invoked on the Dish."
+                "Dish manager did not receive TrackTable."
+                + "Track() command is not invoked on the Dish."
             )
             return (
                 [ResultCode.FAILED],
                 [
-                    "Timeout occurred while waiting to "
-                    + " generate programTrackTable."
+                    "Dish manager did not receive TrackTable. "
+                    + "Track() command is not invoked on the Dish."
                 ],
             )
 
@@ -750,20 +802,32 @@ class Configure(DishLNCommand):
             )
         return [ResultCode.OK], [""]
 
-    def is_tracktable_provided(self) -> bool:
+    def is_tracktable_provided(self) -> str:
         """
         Returns True if programTrackTable is provided to dish.
         """
+        flag = "FALSE"
+
         start_time = time.time()
         elapsed_time = 0
         while elapsed_time < self.component_manager.command_timeout:
+            if self.component_manager.abort_event.is_set():
+                # self.component_manager.abort_event.clear()
+                # self.logger.info("Abort event is cleared")
+                self.logger.info(
+                    "AbortCommands() command is invoked while"
+                    + " configuring dish."
+                )
+                flag = "ABORTED"
+                return flag
             with self.component_manager.tango_operation_execution_lock:
                 track_table = self.dish_master_adapter.programTrackTable
             if len(track_table) > 0:
-                return True
+                flag = "TRUE"
+                return flag
             time.sleep(0.1)
             elapsed_time = time.time() - start_time
-        return False
+        return flag
 
     # pylint: disable=arguments-differ
     def check_device_state(