+++ src/ska_tmc_dishleafnode/manager/component_manager.py
@@ -39,6 +39,7 @@ from ska_tmc_common.lrcr_callback import LRCRCallback
 
 from ska_tmc_dishleafnode.az_el_converter import AzElConverter
 from ska_tmc_dishleafnode.commands import (
+    AbortCommands,
     Configure,
     ConfigureBand,
     EndScan,
@@ -89,7 +90,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
         sleep_time: int = 1,
         dish_availability_check_timeout: int = 40,
         command_timeout: int = 15,
-        is_dish_abort_commands: bool = False,
+        is_dish_abort_commands_enabled: bool = False,
         adapter_timeout: int = 2,
         elevation: float = 0.0,
         azimuth: float = 0.0,
@@ -125,6 +126,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
             proxy_timeout=proxy_timeout,
             sleep_time=sleep_time,
         )
+        self.rlock = threading.RLock()
         self._device = DishDeviceInfo(dish_dev_name)
         self.logger = logger
         __adapter_factory = AdapterFactory()
@@ -154,7 +156,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
         self.elevation_max_limit = elevation_max_limit
         self.elevation_min_limit = elevation_min_limit
         self.el_limit = False
-        self.is_dish_abort_commands = is_dish_abort_commands
+        self.is_dish_abort_commands_enabled = is_dish_abort_commands_enabled
         self.radec_value = ""
         self.process_manager = Manager()
         self._actual_pointing = self.process_manager.list()
@@ -201,6 +203,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
         if _liveliness_probe != LivelinessProbeType.NONE:
             self.start_liveliness_probe(_liveliness_probe)
 
+        self.abort_event = threading.Event()
         self.track_table_scheduler = sched.scheduler(time.time, time.sleep)
         self.track_table_entries: int = track_table_entries
         self.pointing_calculation_period: int = pointing_calculation_period
@@ -956,10 +959,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
             state, False otherwise.
         :rtype: boolean
         """
-        if self.dishMode == DishMode.OPERATE and self.pointingState not in (
-            PointingState.NONE,
-            PointingState.UNKNOWN,
-        ):
+        if self.pointingState in (PointingState.TRACK, PointingState.SLEW):
             return True
 
         raise CommandNotAllowed(
@@ -1123,6 +1123,20 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
             argin,
         )
         return task_status, response
+
+    # pylint: disable=arguments-differ
+    def abort_commands(self) -> Tuple[TaskStatus, str]:
+        """
+        Invokes Abort command on Dish manager.
+        """
+        abort_command = AbortCommands(
+            self,
+            logger=self.logger,
+        )
+        self.abort_event.set()
+        self.logger.debug("Abort event is set.")
+        result_code, message = abort_command.invoke_abort()
+        return result_code, message
 
     def is_trackloadstaticoff_allowed(self: DishLNComponentManager) -> bool:
         """Checks if the command TrackLoadStaticOff is allowed.
@@ -1533,7 +1547,12 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
         :rtype: None
         """
         with self.tango_operation_execution_lock:
-            self.dish_adapter.programTrackTable = program_track_table
+            try:
+                self.dish_adapter.programTrackTable = program_track_table
+            except (tango.DevFailed, Exception) as exception:
+                self.logger.exception(
+                    "Exception while writing tracktable: %s", str(exception)
+                )
         self.logger.debug("ProgramTrackTable: %s", program_track_table)
 
     def track_process(
@@ -1644,7 +1663,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
             self.track_table_process.join()
 
     # pylint: disable=arguments-differ
-    def update_device_ping_failure(
+    def update_exception_for_unresponsiveness(
         self: DishLNComponentManager, device_info: DeviceInfo, exception: str
     ) -> None:
         """Set a device to failed and call the relative callback if available
@@ -1655,26 +1674,20 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
         :type: Exception
         :rtype: None
         """
-        device_info.update_unresponsive(True, exception)
-        with self.lock:
+        with self.rlock:
+            device_info.update_unresponsive(True, exception)
             if self.update_availablity_callback is not None:
                 self.update_availablity_callback(False)
 
-    def update_ping_info(
-        self: DishLNComponentManager, ping: int, device_name: str
-    ) -> None:
+    def update_responsiveness_info(self, device_name: str) -> None:
         """
-        Update a device with the correct ping information.
+        Update a device with the correct availability information.
 
         :param dev_name: name of the device
         :type dev_name: str
-        :param ping: device response time
-        :type ping: int
-        :rtype: None
         """
-        with self.lock:
-            self._device.ping = ping
-            self._device.update_unresponsive(False)
+        with self.rlock:
+            self.get_device().update_unresponsive(False, "")
             if self.update_availablity_callback is not None:
                 self.update_availablity_callback(True)