+++ tests/integration/test_dishln_configure_band_command.py
@@ -32,7 +32,7 @@ def configureband_command(tango_context, dishln_name, group_callback):
 
     group_callback["dishMode"].assert_change_event(
         (DishMode.STANDBY_LP),
-        lookahead=2,
+        lookahead=7,
     )
 
     result_fp, unique_id_fp = dish_leaf_node.SetStandbyFPMode()
@@ -45,11 +45,11 @@ def configureband_command(tango_context, dishln_name, group_callback):
     )
     group_callback["longRunningCommandResult"].assert_change_event(
         (unique_id_fp[0], COMMAND_COMPLETED),
-        lookahead=2,
+        lookahead=7,
     )
     group_callback["dishMode"].assert_change_event(
         (DishMode.STANDBY_FP),
-        lookahead=2,
+        lookahead=7,
     )
 
     result_op, unique_id_op = dish_leaf_node.ConfigureBand("1")
@@ -58,7 +58,7 @@ def configureband_command(tango_context, dishln_name, group_callback):
 
     group_callback["longRunningCommandResult"].assert_change_event(
         (unique_id_op[0], COMMAND_COMPLETED),
-        lookahead=2,
+        lookahead=7,
     )
     assert dish_master.configuredBand == Band.B1
 
@@ -68,7 +68,7 @@ def configureband_command(tango_context, dishln_name, group_callback):
     tear_down(dish_leaf_node, dish_master, group_callback)
 
 
-@pytest.mark.sah1589
+# @pytest.mark.sah1589
 @pytest.mark.post_deployment
 @pytest.mark.SKA_mid
 def test_configureband_command(tango_context, group_callback):
@@ -148,7 +148,7 @@ def configureband_command_error_propogation(
     tear_down(dish_leaf_node, dish_master, group_callback)
 
 
-@pytest.mark.sah1589
+# @pytest.mark.sah1589
 @pytest.mark.post_deployment
 @pytest.mark.SKA_mid
 def test_configureband_command_error_propogation(